Make recalc from csv work

+++ app/Helpers/RecalculateBuyingLabels.php
@@ -17,7 +17,7 @@ class RecalculateBuyingLabels
         OrderPayment::where('created_at' , '>=', now()->firstOfMonth())
             ->whereHas('order', function ($q) {$q->whereHas('items.product', function ($q) {$q->where('variation_group', 'styropiany');});})->where('operation_type', 'Wpłata/wypłata bankowa')->get()->unique('order')->pluck('order.id');
 
-        if (empty(BuyingInvoice::where('order_id', $order->id)->first()) && !$order->labels->contains('id', 65)) {
+        if (empty(BuyingInvoice::where('order_id', $order->id)->first()) && !$order->labels->contains('id', 65) && !$order->payments->where('operation_type', 'Wpłata/wypłata bankowa - związana z fakturą zakupową')->first()) {
             $arr = [];
 
             RemoveLabelService::removeLabels($order, [263], $arr , [], auth()->id());
